python-code/python01/02章-AI学习/03_AI_Partner/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile requirements.in
#
altair==6.0.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   openai
attrs==26.1.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==7.0.5
    # via streamlit
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.7
    # via requests
click==8.3.2
    # via streamlit
colorama==0.4.6
    # via
    #   click
    #   tqdm
distro==1.9.0
    # via openai
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via streamlit
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jiter==0.14.0
    # via openai
jsonschema==4.26.0
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.3
    # via jinja2
narwhals==2.19.0
    # via altair
numpy==2.4.4
    # via
    #   pandas
    #   pydeck
    #   streamlit
openai==2.31.0
    # via -r requirements.in
packaging==26.0
    # via
    #   altair
    #   streamlit
pandas==3.0.2
    # via streamlit
pillow==12.2.0
    # via streamlit
protobuf==7.34.1
    # via streamlit
pyarrow==23.0.1
    # via streamlit
pydantic==2.12.5
    # via openai
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.2
    # via -r requirements.in
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.33.1
    # via streamlit
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.3
    # via gitdb
sniffio==1.3.1
    # via openai
streamlit==1.56.0
    # via -r requirements.in
tenacity==9.1.4
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.5
    # via streamlit
tqdm==4.67.3
    # via openai
typing-extensions==4.15.0
    # via
    #   altair
    #   openai
    #   pydantic
    #   pydantic-core
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2026.1
    # via pandas
urllib3==2.6.3
    # via requests
watchdog==6.0.0
    # via streamlit
